Add files via upload
This data set contains all ranked games as of the date of collection from the BGG database. Unranked games are ignored as they have not been rated by enough BGG users (a game should receive at least 30 votes to be eligible for ranking) which render the quality of games unsatisfactory to be considered for the analysis.